# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QA Automation Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: environment
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortDir
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get project stats (audit count, bug count, avg score)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"
...
  /projects/{id}/tech-profile:
    get:
      operationId: getProjectTechProfile
      tags: [projects]
      summary: Get technology profile for a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technology profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechProfile"
    put:
      operationId: updateProjectTechProfile
      tags: [projects]
      summary: Update technology profile for a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechProfile"
      responses:
        "200":
          description: Updated technology profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechProfile"

  # Audits
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: auditRunId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a PDF report for an audit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # Scheduled Audits
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TechProfile:
      type: object
      properties:
        framework:
          type: ["string", "null"]
        frontendStack:
          type: ["string", "null"]
        backendStack:
          type: ["string", "null"]
        cms:
          type: ["string", "null"]
        database:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Project:
      type: object
      required: [id, name, url, environment, createdAt, createdById]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        environment:
          type: string
          enum: [development, staging, production]
        description:
          type: ["string", "null"]
        auditTemplate:
          type: string
          enum: [react, vite, nextjs, wordpress, shopify, laravel, nodejs_api, static, custom]
        techProfile:
          type: ["object", "null"]
        createdById:
          type: integer
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastAuditAt:
          type: ["string", "null"]
        auditCount:
          type: integer
        openBugCount:
          type: integer
    ProjectInput:
      type: object
      required: [name, url, environment]
      properties:
        name:
          type: string
        url:
          type: string
        environment:
          type: string
          enum: [development, staging, production]
        description:
          type: string
        auditTemplate:
          type: string
          enum: [react, vite, nextjs, wordpress, shopify, laravel, nodejs_api, static, custom]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        environment:
          type: string
          enum: [development, staging, production]
        description:
          type: string
        auditTemplate:
          type: string
          enum: [react, vite, nextjs, wordpress, shopify, laravel, nodejs_api, static, custom]
    ProjectStats:
      type: object
      required: [projectId, totalAudits, totalBugs, openBugs, avgPerformanceScore, avgAccessibilityScore, avgSeoScore, avgBestPracticesScore]
      properties:
        projectId:
          type: integer
        totalAudits:
          type: integer
        totalBugs:
          type: integer
        openBugs:
          type: integer
        avgPerformanceScore:
          type: ["number", "null"]
        avgAccessibilityScore:
          type: ["number", "null"]
        avgSeoScore:
          type: ["number", "null"]
        avgBestPracticesScore:
          type: ["number", "null"]
...
    Report:
      type: object
      required: [id, projectId, auditRunId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        auditRunId:
          type: integer
        status:
          type: string
          enum: [generating, ready, failed]
        fileUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReportInput:
      type: object
      required: [auditRunId]
      properties:
        auditRunId:
          type: integer

    # Scheduled Audits
...